feat: renaming model
